refactor(diagnostics): extract shared CLI/DB health checks to internal/sys
- Create internal/sys/diagnostics.go with CheckCliAvailable() and
  CheckDatabaseHealth() shared by both admin handler and doctor command
- Cache CLI version in Handler struct at construction time
- Remove duplicate getCliVersion() and checkDatabaseHealth() from handler.go
- Update handler_test.go and doctor.go to use sys package
- Fix errCh channel to use ErrChan() correctly in server.Start()
